split simple_train into its own module
split both mlx training and linux_train into a simple_train module.

This is the foundation for future API handlers and multi processing work.

It also clears up the train.py file a lot and makes the code a lot more readable
